Make host and metadata updatable in webview Fixes #12
