+++ src/extension.ts
@@ -4,7 +4,7 @@ import { AdressList as HostsTreeView } from "./hosts/list";
 import { MetasList } from "./metas/list";
 import { ProtosTree as ProtosTreeView } from "./protos/tree";
 import { Storage } from "./storage/storage";
-import { Input } from "./webview/input";
+import { Input } from "./webview/data";
 import { WebViewFactory } from "./webview/panel";
 
 export function activate(context: vscode.ExtensionContext) {
