--- src/webview/response.ts
@@ -1,6 +1,0 @@
-export class Response {
-  public rez: boolean;
-  constructor(public output: string) {
-    this.rez = true;
-  }
-}
